# Cache language menu