--- a/app/Libraries/LayoutCache.php
+++ b/app/Libraries/LayoutCache.php
@@ -3,13 +3,15 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
+declare(strict_types=1);
+
 namespace App\Libraries;
 
 class LayoutCache
 {
     private array $cache = [];
 
-    public function getBeatmapsetFilters()
+    public function getBeatmapsetFilters(): string
     {
         $key = __FUNCTION__.':'.app()->getLocale();