… resources/views/layout/header_mobile/nav.blade.php