The full header cache will take a bit longer to do.

- [x] depends on #9343 as it uses the same layout cache thing

Before:
```
Run 1:   2153 requests in 10.08s, 637.05MB read
Run 2:   2195 requests in 10.08s, 649.48MB read
Run 3:   2167 requests in 10.08s, 641.20MB read
Run 4:   2204 requests in 10.08s, 652.15MB read
Run 5:   2190 requests in 10.09s, 648.00MB read
Run 6:   2143 requests in 10.08s, 634.09MB read
Run 7:   2110 requests in 10.08s, 624.33MB read
Run 8:   2205 requests in 10.17s, 652.69MB read
Run 9:   2158 requests in 10.08s, 638.53MB read
Run 10:   2162 requests in 10.08s, 639.72MB read
Final run:
Running 10s test @ https://on.nanaya.pro/beatmapsets/
  4 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    55.53ms   12.90ms 377.68ms   95.84%
    Req/Sec    54.10      8.09    70.00     87.75%
  2179 requests in 10.07s, 644.98MB read
Requests/sec:    216.28
Transfer/sec:     64.02MB
```

After:
```
Run 1:   3137 requests in 10.06s, 0.88GB read
Run 2:   3060 requests in 10.06s, 0.86GB read
Run 3:   3232 requests in 10.07s, 0.91GB read
Run 4:   3123 requests in 10.06s, 0.88GB read
Run 5:   3143 requests in 10.06s, 0.88GB read
Run 6:   3178 requests in 10.05s, 0.89GB read
Run 7:   3098 requests in 10.07s, 0.87GB read
Run 8:   3121 requests in 10.15s, 0.88GB read
Run 9:   3162 requests in 10.06s, 0.89GB read
Run 10:   3153 requests in 9.97s, 0.89GB read
Final run:
Running 10s test @ https://on.nanaya.pro/beatmapsets/
  4 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    41.01ms   30.50ms 420.26ms   97.90%
    Req/Sec    78.68     11.50   120.00     86.32%
  3172 requests in 10.15s, 0.89GB read
Requests/sec:    312.49
Transfer/sec:     90.08MB
```