@notbakaneko 2022-10-05
32.32req/s to 69.54req/s ....